+++ frontend/src/components/expense/ExpenseItem/ExpenseItem.style.ts
@@ -3,7 +3,7 @@ import { css } from '@emotion/react';
 import { Theme } from 'hang-log-design-system';
 
 export const containerStyling = css({
-  width: '80%',
+  width: '100%',
 
   '& p': {
     color: Theme.color.gray600,
