+++ frontend/src/stories/expense/ExpenseInformation.stories.tsx
@@ -2,18 +2,21 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import ExpenseInformation from '@components/expense/ExpenseInformation/ExpenseInformation';
 
-import { expense } from '@mocks/data/expense';
+import { useExpenseQuery } from '@hooks/api/useExpenseQuery';
 
 const meta = {
   title: 'expense/ExpenseInformation',
   component: ExpenseInformation,
   args: {
     tripId: 1,
-    title: expense.title,
-    startDate: expense.startDate,
-    endDate: expense.endDate,
-    cities: expense.cities,
   },
+  decorators: [
+    (Story) => {
+      useExpenseQuery(1);
+
+      return <Story />;
+    },
+  ],
 } satisfies Meta<typeof ExpenseInformation>;
 
 export default meta;
